+++ ios/TimeSenseWidget/TimeSenseWidgetBundle.swift
@@ -1,0 +1,11 @@
+import WidgetKit
+import SwiftUI
+
+@main
+struct TimeSenseWidgetBundle: WidgetBundle {
+    var body: some Widget {
+        UsableTimeWidget()
+        NextEventWidget()
+        BestNextActionWidget()
+    }
+}